Refactor admin applications management page
- Consolidated applications management functionality into a single page.
- Enhanced UI with improved statistics cards and filter tabs for application statuses.
- Implemented application status updates and deletion features.
- Removed the previous applications page to streamline the admin interface.
- Updated sidebar navigation to reflect the new structure.
